[CO-327] Use Example typeclass
1. Import and use `Example` typeclass in `../V1/Types.hs`
instead of `Arbitrary`
2. Resolve circular dependencies
`Errors.hs` -> `Types.hs` -> `Example.hs` -> `Response.hs` -> `Errors.hs`
by moving `Example` instances from `Example.hs` to where the datatypes
are defined, so that `Example.hs` does not need to import `Response.hs`
or `Types.hs`.
