+++ wallet-new/src/Cardano/Wallet/API/Response.hs
@@ -16,12 +16,13 @@ import           Prelude
 import           Universum (Buildable, decodeUtf8, toText, (<>))
 
 import           Cardano.Wallet.API.Response.JSend (ResponseStatus (..))
+import           Cardano.Wallet.API.V1.Swagger.Example (Example, example)
 import           Control.Lens
 import           Data.Aeson
 import           Data.Aeson.Encode.Pretty (encodePretty)
 import           Data.Aeson.TH
 import qualified Data.Char as Char
-import           Data.Swagger as S
+import           Data.Swagger as S hiding (Example, example)
 import qualified Data.Text.Buildable
 import           Data.Typeable
 import           Formatting (bprint, build, (%))
@@ -158,6 +159,12 @@ paginate PaginationParams{..} rawResultSet =
                                }
         slice                  = take pp . drop ((cp - 1) * pp)
     in (slice rawResultSet, metadata)
+
+instance Example Metadata
+instance Example a => Example (WalletResponse a) where
+    example = WalletResponse <$> example
+                             <*> pure SuccessStatus
+                             <*> example
 
 
 -- | Creates a 'WalletResponse' with just a single record into it.
